fx_v46/requirements.txt
MetaTrader5>=5.0
numpy>=1.26
pandas>=2.2
python-dotenv>=1.0
requests>=2.32
fastapi>=0.100
uvicorn>=0.22
jinja2>=3.1